-setup for graphql submissions
