+++ apollo-server/definitions/todo.md
@@ -1,0 +1,7 @@
+Unimplemented Challenge resolvers:
+
+# createChallengeGroup(data: CreateChallengeGroupInput!): ChallengeGroup!
+
+# updateChallengeGroup(challengeGroupId: ID!, data: UpdateChallengeGroupInput!): ChallengeGroup!
+
+# deleteChallenge(challengeId: ID!): Challenge!
