+++ apollo-server/resolvers/challenge.js
@@ -67,19 +67,6 @@ function myChallenges(parent, args, { id }) {
   return getMyChallenges(id);
 }
 
-/**
- * returns all of user's submissions for a specific challenge
- * @param {*} userid
- * @param {*} challengid
- */
-function submissions(userid, challengid) {}
-
-/**
- * Returns a specific submission
- * @param {*} submissionId
- */
-function submission() {}
-
 // Mutations:
 
 /**
@@ -102,15 +89,6 @@ const Challenge = {
   },
   // submissions,
 };
-/**
- * creates a new submission
- */
-function createSubmission() {}
-
-/**
- * Updates a submission
- */
-function updateSubmission() {}
 
 module.exports = {
   challengeGroups,
