+++ apollo-server/resolvers/submissions.js
@@ -1,0 +1,25 @@
+// QUERIES:
+
+/**
+ * returns all of user's submissions for a specific challenge
+ * @param {*} userid
+ * @param {*} challengid
+ */
+function submissions(userid, challengid) {}
+
+/**
+ * Returns a specific submission
+ * @param {*} submissionId
+ */
+function submission() {}
+
+// MUTATIONS:
+/**
+ * creates a new submission
+ */
+function createSubmission() {}
+
+/**
+ * Updates a submission
+ */
+function updateSubmission() {}
