+++ postgresDb/initialization/submissionsInit.js
@@ -1,0 +1,4 @@
+require('dotenv').config();
+const db = require('../index');
+const { makeQuery } = require('../pgHelpers');
+
