+++ postgresDb/submissions/submissionsHelpers.js
@@ -1,0 +1,7 @@
+const db = require('../index');
+const {
+  makeQuery,
+  makeInsert,
+  cleanProps,
+  renameProp,
+} = require('../pgHelpers');
